% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualize.R
\name{annualize}
\alias{annualize}
\title{Sum a GHCN record over a set of months to generate annual estimates.}
\usage{
annualize(station, months = 1:12)
}
\arguments{
\item{station}{GHCN station data}

\item{months}{The months over which to annualize}
}
\value{
A \link{data.frame} of annualized data.
}
\description{
Sum a GHCN record over a set of months to generate annual estimates.
}
