% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.new.R
\name{arrow.new}
\alias{arrow.new}
\title{A north arrow.}
\usage{
arrow.new(x, y, height, width, adj)
}
\arguments{
\item{x}{The x position of the arrow}

\item{y}{The y position of the arrow}

\item{height}{The height of the arrow}

\item{width}{The width of the arrow}

\item{adj}{The adjustment for the arrow}
}
\description{
A north arrow.
}
