% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGapCutoff.R
\name{calcGapCutoff}
\alias{calcGapCutoff}
\title{Calculates the empirical cumulative distribution function.}
\usage{
calcGapCutoff(rleVector, pLevel = 0.95)
}
\arguments{
\item{rleVector}{the RLE vector}

\item{pLevel}{a p level}
}
\value{
The cutoff value
}
\description{
This function calculates the empirical cumulative distribution function,
then uses a user-defined probability level to calculate the "normal" gap length.
That way, GHCN records (years) with unusually long gap lengths may be discarded.
}
