% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDD.R
\name{calcGDD}
\alias{calcGDD}
\title{A function to calculate GDD, given daily maximum and minimum temperature series,
and maximum and minimum growth temperatures.}
\usage{
calcGDD(tmin.vector, tmax.vector, t.base, t.cap = NULL)
}
\arguments{
\item{tmin.vector}{A vector of daily tmin values}

\item{tmax.vector}{A vector of daily tmax values}

\item{t.base}{A base temperature}

\item{t.cap}{An (optional) cap temperature}
}
\value{
A vector of daily GDD values
}
\description{
GDD is calculated as
$((TMINc+TMAXc)/2)-TBASE$
Where TMINc and TMAXc are measured temperatures bounded by TBASE as a minimum
and perhaps a TCAP as a maximum.
}
