% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_plotter.R
\name{climate_plotter}
\alias{climate_plotter}
\title{Plot climatology for a GHCN station and element (variable).}
\usage{
climate_plotter(data, station, element)
}
\arguments{
\item{element}{}
}
\description{
Plot climatology for a GHCN station and element (variable).
}
