% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distill.R
\name{distill}
\alias{distill}
\title{Use ghostscript to distill a PDF, possibly to grayscale.}
\usage{
distill(file, gray = F)
}
\arguments{
\item{file}{A file name}

\item{gray}{Whether the output should be grayscale}
}
\description{
Use ghostscript to distill a PDF, possibly to grayscale.
}
