% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filledContour.R
\name{filledContourAdd}
\alias{filledContourAdd}
\title{Add a filled contour to a plot.}
\usage{
filledContourAdd(x, y = 1, maxpixels = 1e+05, levels, col)
}
\arguments{
\item{col}{}
}
\description{
Add a filled contour to a plot.
}
