% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_restricted_site_data.R
\name{get_restricted_site_data}
\alias{get_restricted_site_data}
\title{Get the restricted archaeological site and chronometric data from tDAR}
\usage{
get_restricted_site_data(tdar_user = Sys.getenv("tdar_un"),
  tdar_password = Sys.getenv("tdar_pw"))
}
\arguments{
\item{tdar_user}{Your tDAR user name}

\item{tdar_password}{Your tDAR password}
}
\value{
A data.frame of site and chronometric data
}
\description{
Get the restricted archaeological site and chronometric data from tDAR
}
