% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_restricted_vegetation_atlas_data.R
\name{get_restricted_vegetation_atlas_data}
\alias{get_restricted_vegetation_atlas_data}
\title{Get the restricted vegetation_atlas data from tDAR}
\usage{
get_restricted_vegetation_atlas_data(tdar_user = Sys.getenv("tdar_un"),
  tdar_password = Sys.getenv("tdar_pw"))
}
\arguments{
\item{tdar_user}{Your tDAR user name}

\item{tdar_password}{Your tDAR password}
}
\value{
A Simple Features Data Frame of vegetation atlas data
}
\description{
Get the restricted vegetation_atlas data from tDAR
}
