% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMissingRLE.R
\name{getMissingRLE}
\alias{getMissingRLE}
\title{Get the run length encoding of joint NAs in the climate records.}
\usage{
getMissingRLE(vect)
}
\arguments{
\item{vect}{A numebric vector}
}
\value{
the RLE of the vector
}
\description{
Useful for deciding the cutoff value for the number of contiguous days.
}
