% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcnCleaner.R
\name{ghcnCleaner}
\alias{ghcnCleaner}
\title{Cleans ghcn data by four iterated processes.}
\usage{
ghcnCleaner(data.list, year.range, na.cutoff, min.years = length(year.range))
}
\arguments{
\item{min.years}{}
}
\description{
(1) crop to year.range
(2) purge years with gaps in the data longer than the cutoff length
(3) interpolate across gaps <= the cutoff length
(4) fill by average those days at the beginning and end of years with missing data
(5) remove stations with less than min.years years of complete data
}
