% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ghcn.R
\name{prepare_ghcn}
\alias{prepare_ghcn}
\title{Download and clean daily climate records from the Global Historical Climate Database.}
\usage{
prepare_ghcn(region, label, calibration.years, google_maps_elevation_api_key,
  raw_dir = "./data/raw_data/ghcn/",
  derived_dir = "./data/derived_data/ghcn/", force.redo = FALSE)
}
\arguments{
\item{region}{A spatial object defining the region}

\item{label}{A label for the region}

\item{calibration.years}{The years for defining climatology}

\item{google_maps_elevation_api_key}{the users Google Maps Elevation API key}

\item{raw_dir}{a directory in which to store the raw GHCN downloads}

\item{derived_dir}{a directory in which to store the processed GHCN output}

\item{force.redo}{Should the GHCN data be re-processed?}
}
\value{
A clean set of GHCN data
}
\description{
Download and clean daily climate records from the Global Historical Climate Database.
}
