% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_marcott.R
\name{prepare_marcott}
\alias{prepare_marcott}
\title{Download and prepare data from Mann et al. 2008 and Marcott et al. 2013.}
\usage{
prepare_marcott(calibration.years, raw_dir = "./data/raw_data/",
  derived_dir = "./data/derived_data/")
}
\arguments{
\item{calibration.years}{The years used to define climatology}

\item{raw_dir}{a directory in which to store the raw downloads}

\item{derived_dir}{a directory in which to store the processed output}
}
\value{
the Marcott dataset
}
\description{
Download and prepare data from Mann et al. 2008 and Marcott et al. 2013.
}
