% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewrapRows.R
\name{rewrapRows}
\alias{rewrapRows}
\title{A function that rewraps a vector into a matrix.}
\usage{
rewrapRows(vec, n)
}
\arguments{
\item{vec}{A vector}

\item{n}{the length of each row.}
}
\value{
A matrix
}
\description{
User must provide a vector (n) specifying the length of rows.
n can be either a constant (in which case it will be repeated), or a vector.
}
