% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalebar.new.R
\name{scalebar.new}
\alias{scalebar.new}
\title{A simple scale bar.}
\usage{
scalebar.new(d, xy = NULL, height = NULL, line.offset = c(0, 0),
  side = "right", lab.side = "top", col = "black", lonlat = NULL, label,
  adj = c(0.5, -0.5), lwd = 2, ...)
}
\arguments{
\item{...}{}
}
\description{
A simple scale bar.
}
