% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_time_plot.R
\name{space_time_plot}
\alias{space_time_plot}
\title{A space-time plot.}
\usage{
space_time_plot(the_brick, the_brick_upper = NULL, the_brick_lower = NULL,
  out_file, title = NULL, time, timelim, timeaxis, timelab = "Year AD",
  zbreaks = NULL, zlab, zaxis, zcolors = rev(brewer.pal(11, "RdYlBu")),
  fig_width = 6.5, graph_height = 1.5, margin = 0.1, pt_size = 8,
  smooth = FALSE, extra_plot_fun = NULL, extra_legend_fun = NULL)
}
\arguments{
\item{extra_legend_fun}{}
}
\description{
Given a raster brick (and possibly uncertainties/errors),
plot a map of the average across time, and a
graph of the average across space.
}
