% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDailyMeanSD.R
\name{calcDailyMeanSD}
\alias{calcDailyMeanSD}
\title{This function smooths periodic (annual) data in the GHCN format.}
\usage{
calcDailyMeanSD(daily.data, unwrapped = F, plot = F)
}
\arguments{
\item{daily.data}{GHCN data}

\item{unwrapped}{Whether the data are already unwrapped}

\item{plot}{Whether to create a polar plot of the climatology.}
}
\value{
A list containing the day-of-year, mean, and sd.
}
\description{
First, it converts all data to DOY data (366-day-year).
Then, it calculates a periodic generalized additive model
from the data using cyclic cubic regression splines for smoothing.
It then calculates the daily-wise standard deviations of the residuals,
and again smooths using the same type of GAM.
Optionally, you may plot the results.
}
