% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDD.R
\name{calcGDD}
\alias{calcGDD}
\title{A function to calculate GDD, given daily maximum and minimum temperature series,
and maximum and minimum growth temperatures.}
\usage{
calcGDD(tmin.vector, tmax.vector, t.base, t.cap = NULL)
}
\arguments{
\item{t.cap}{}
}
\description{
GDD is calculated as
$\link{(TMINc+TMAXc)/2}-TBASE$
Where TMINc and TMAXc are measured temperatures bounded by TBASE as a minimum
and perhaps a TCAP as a maximum.
}
