% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_plotter.R
\name{climate_plotter}
\alias{climate_plotter}
\title{Plot climatology for a GHCN station and element (variable).}
\usage{
climate_plotter(data, station, element)
}
\arguments{
\item{data}{GHCN data}

\item{station}{A station name}

\item{element}{An element (variable) name}
}
\value{
A ggplot2 object
}
\description{
Plot climatology for a GHCN station and element (variable).
}
