% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filledContour.R
\name{filledContourAdd}
\alias{filledContourAdd}
\title{Add a filled contour to a plot.}
\usage{
filledContourAdd(x, y = 1, maxpixels = 1e+05, levels, col)
}
\arguments{
\item{x}{a raster object}

\item{y}{Integer. The layer number of x (if x has multiple layers)}

\item{maxpixels}{the maximum number of pixels to plot}

\item{levels}{The levels to plot}

\item{col}{a vector of colors}
}
\value{
An additional plot on the current graphics device
}
\description{
Add a filled contour to a plot.
}
