% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghcnCleaner.R
\name{ghcnCleaner}
\alias{ghcnCleaner}
\title{Cleans ghcn data by four iterated processes.}
\usage{
ghcnCleaner(data.list, year.range, na.cutoff, min.years = length(year.range))
}
\arguments{
\item{data.list}{a list of GHCN data sets}

\item{year.range}{the intended years}

\item{na.cutoff}{the maximum length of gaps, above which a dataset is excluded.}

\item{min.years}{the minimum number of years a dataset must include, below which a dataset is excluded.}
}
\value{
a clean set of ghcn data
}
\description{
(1) crop to year.range
(2) purge years with gaps in the data longer than the cutoff length
(3) interpolate across gaps <= the cutoff length
(4) fill by average those days at the beginning and end of years with missing data
(5) remove stations with less than min.years years of complete data
}
