% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ghcn.R
\name{prepare_ghcn}
\alias{prepare_ghcn}
\title{Download and clean daily climate records from the Global Historical Climate Database.}
\usage{
prepare_ghcn(region, label, calibration.years, google_maps_elevation_api_key,
  force.redo = FALSE)
}
\arguments{
\item{force.redo}{}
}
\description{
Download and clean daily climate records from the Global Historical Climate Database.
}
