% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_marcott.R
\name{prepare_marcott}
\alias{prepare_marcott}
\title{Download and prepare data from Mann et al. 2008 and Marcott et al. 2013.}
\usage{
prepare_marcott(calibration.years)
}
\arguments{
\item{calibration.years}{The years used to define climatology}
}
\description{
Download and prepare data from Mann et al. 2008 and Marcott et al. 2013.
}
