% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalebar.new.R
\name{scalebar.new}
\alias{scalebar.new}
\title{A simple scale bar.}
\usage{
scalebar.new(d, xy = NULL, height = NULL, line.offset = c(0, 0),
  side = "right", lab.side = "top", col = "black", lonlat = NULL, label,
  adj = c(0.5, -0.5), lwd = 2, ...)
}
\arguments{
\item{d}{the length of the scalebar in meters}

\item{xy}{the position of the scale bar}

\item{height}{the height of the scale bar in inches}

\item{line.offset}{the offset of the position of the scale bar}

\item{side}{which side of the plot the scale bar should be on}

\item{lab.side}{which side of the scale bar the label should be on}

\item{col}{the color of the scale bar}

\item{lonlat}{is the plot in georgaphic coordinates?}

\item{label}{The label}

\item{adj}{An adjustment for the scale bar}

\item{lwd}{the line width of the scale bar}

\item{...}{Additional parameters passed on to text}
}
\value{
A scalebar is added to the current graphics device.
}
\description{
A simple scale bar.
}
