% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdModulator.R
\name{sdModulator}
\alias{sdModulator}
\title{Modulate a record by a deviation of its standard score.}
\usage{
sdModulator(data.df, temp.change.sd, t.base, t.cap = NULL)
}
\arguments{
\item{t.cap}{}
}
\description{
Modulate a record by a deviation of its standard score.
}
