% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_time_video.R
\name{space_time_video}
\alias{space_time_video}
\title{A space-time video}
\usage{
space_time_video(the_brick, the_brick_upper = NULL, the_brick_lower = NULL,
  out_file, line = 75, title = NULL, time, timelim, timeaxis,
  timelab = "Year AD", zbreaks = NULL, zlab, zaxis,
  zcolors = rev(RColorBrewer::brewer.pal(11, "RdYlBu")), fig_width = 6.5,
  graph_height = 1.5, margin = 0.1, pt_size = 8, smooth = FALSE,
  length = 60, extra_plot_fun = NULL, extra_legend_fun = NULL)
}
\arguments{
\item{extra_legend_fun}{}
}
\description{
Given a raster brick (and possibly uncertainties/errors),
plot a video of the brick through time, and a
graph of the average across space, with a bar indicating year.
}
\keyword{internal}
