% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrapRows.R
\name{unwrapRows}
\alias{unwrapRows}
\title{A function that unwraps a matrix and only keeps the first n elements.}
\usage{
unwrapRows(mat, n)
}
\arguments{
\item{n}{}
}
\description{
n can be either a constant (in which case it will be repeated), or a vector.
}
